+++ src/internal/constants.ts
@@ -38,7 +38,12 @@ export enum DecoratorKeys {
    * Storage location for Virtual Populates
    * -> Use only for a class
    */
-  VirtualPopulate = 'typegoose:virtualPopulate'
+  VirtualPopulate = 'typegoose:virtualPopulate',
+  /**
+   * Storage location for Query Methods
+   * -> Use only for a class
+   */
+  QueryMethod = 'typegoose:queryMethod'
 }
 
 /** This Enum is meant for baseProp to decide for diffrent props (like if it is an arrayProp or prop or mapProp) */