+++ github-page/changelog.md
@@ -37,6 +37,8 @@ redirect_from:
 - Add PropOption `addNullToEnum`
 - Remove Deprecated value `overwrite` for `VirtualOptions`
 - Remove instance properties from Model type (remove `& T` from `ModelType`)
+- Add class decorator `queryMethod`
+- [IC] rename file `optionsProp` to `modelOptions`
 - [IC] Replace mocha & chai with jest
 - [IC] Completly remove `TG_USE_NEW_ENUM` from documentation & code
 - [IC] Replace almost all if-throw with the internal `assertion` function